fix sign-compare warnings

+++ tests/Cases/Common/ColumnTypes/TestInt.cpp
@@ -8,7 +8,7 @@ TEST(TestInt, getset) {
 	ASSERT_EQ(value.get(), 0x7fff0000);
 
 	value = cql::Int(-0x80000000);
-	ASSERT_EQ(value.get(), -0x80000000);
+	ASSERT_EQ(value.get(), static_cast<std::int32_t>(-0x80000000));
 }
 
 TEST(TestInt, encodeBody) {
